Refactor all User cells to a single file
